+++ app/upgrades/v1/data_params_test.go
@@ -3,6 +3,7 @@ package v1_test
 import (
 	"testing"
 
+	wasmtypes "github.com/CosmWasm/wasmd/x/wasm/types"
 	"github.com/stretchr/testify/require"
 
 	"github.com/babylonlabs-io/babylon/app"
@@ -12,17 +13,28 @@ import (
 func TestHardCodedBtcStakingParamsAreValid(t *testing.T) {
 	bbnApp := app.NewTmpBabylonApp()
 	for _, upgradeData := range UpgradeV1Data {
-		loadedParamas, err := v1.LoadBtcStakingParamsFromData(bbnApp.AppCodec(), upgradeData.BtcStakingParamStr)
+		params, err := v1.LoadBtcStakingParamsFromData(bbnApp.AppCodec(), upgradeData.BtcStakingParamStr)
 		require.NoError(t, err)
-		require.NoError(t, loadedParamas.Validate())
+		require.NoError(t, params.Validate())
 	}
 }
 
 func TestHardCodedFinalityParamsAreValid(t *testing.T) {
 	bbnApp := app.NewTmpBabylonApp()
 	for _, upgradeData := range UpgradeV1Data {
-		loadedParamas, err := v1.LoadFinalityParamsFromData(bbnApp.AppCodec(), upgradeData.FinalityParamStr)
+		params, err := v1.LoadFinalityParamsFromData(bbnApp.AppCodec(), upgradeData.FinalityParamStr)
 		require.NoError(t, err)
-		require.NoError(t, loadedParamas.Validate())
+		require.NoError(t, params.Validate())
+	}
+}
+
+func TestHardCodedWasmParamsAreValid(t *testing.T) {
+	bbnApp := app.NewTmpBabylonApp()
+
+	for _, upgradeData := range UpgradeV1Data {
+		params, err := v1.LoadCosmWasmParamsFromData(bbnApp.AppCodec(), upgradeData.CosmWasmParamStr)
+		require.NoError(t, err)
+		require.NotNil(t, params)
+		require.Equal(t, params.InstantiateDefaultPermission, wasmtypes.AccessTypeEverybody)
 	}
 }