+++ app/upgrades/v1/testnet/cosmwasm_params.go
@@ -1,0 +1,10 @@
+package testnet
+
+// CosmWasm parameters for testnet should allow everybody to
+// upload and instantiate.
+const CosmWasmParamStr = `{
+  "code_upload_access": {
+    "permission": "Everybody"
+  },
+  "instantiate_default_permission": "Everybody"
+}`