+++ CHANGELOG.md
@@ -37,14 +37,26 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Bug Fixes
+
+* [#140](https://github.com/babylonlabs-io/babylon/pull/140) Removed `unbonding`
+and add `verified` to delegation status parse `NewBTCDelegationStatusFromString`.
+
 ### State Machine Breaking
+
+* [#132](https://github.com/babylonlabs-io/babylon/pull/132) Add CosmWasm parameters
+update during v1 upgrade handler.
+
+### Misc Improvements
 
 * [#130](https://github.com/babylonlabs-io/babylon/pull/130) Fix bugs in the
 transaction fee refunding mechanism for covenant signatures and finality signatures
 * [#125](https://github.com/babylonlabs-io/babylon/pull/125) Implement ADR-028 and
 refund transaction fee for certain transactions from protocol stakeholders
+* [#137](https://github.com/babylonlabs-io/babylon/pull/137) Adapt tests to the
+pre-approval flow.
 * [#138](https://github.com/babylonlabs-io/babylon/pull/138) Intercept staking module
-messages inside `authz.MsgExec`
+messages inside `authz.MsgExec`.
 
 ### Improvements
 
@@ -52,6 +64,11 @@ messages inside `authz.MsgExec`
 * [#131](https://github.com/babylonlabs-io/babylon/pull/131) Add new staking events
 * [#113](https://github.com/babylonlabs-io/babylon/pull/113) Add multibuild binary
 for upgrade handler `testnet` and `mainnet`.
+
+### Bug Fixes
+
+* [#141](https://github.com/babylonlabs-io/babylon/pull/141) Generate voting
+power events only once when reaching covenant committee quorum
 
 ## v0.11.0